+++ .classpath
@@ -7,6 +7,7 @@
 		</attributes>
 	</classpathentry>
 	<classpathentry kind="con" path="org.eclipse.jdt.junit.JUNIT_CONTAINER/5"/>
-	<classpathentry kind="lib" path="J:/Cs 320/320-F21-Track-2/lib/snakeyaml-1.29.jar"/>
+	<classpathentry kind="lib" path="./lib/snakeyaml-1.29.jar"/>
+	<classpathentry kind="src" path="RuleEngine"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
