+++ .gitignore
@@ -24,3 +24,6 @@ hs_err_pid*
 
 # Project File
 .project
+
+# Bin
+bin
