+++ RuleAI/oracles/Action.java
@@ -1,6 +1,7 @@
+package oracles;
+
 // Any Game Input / Decision
 // Placeholder for GameButton object
-
 public class Action extends GameObject {
     public Action() {
         super();
