+++ RuleAI/oracles/DepthOracle.java
@@ -1,3 +1,4 @@
+package oracles;
 import java.util.ArrayList;
 import java.util.Deque;
