+++ RuleAI/oracles/RandomOracle.java
@@ -1,3 +1,4 @@
+package oracles;
 import java.util.ArrayList;
 import java.util.Deque;
