+++ RuleEngine/nodes/ALNode.java
@@ -1,5 +1,7 @@
+package nodes;
 import java.util.ArrayList;
 import java.lang.reflect.Field;
+import engine.GameState;
 
 // Arithmetic-Logic Node. Performs basic arithmetic and logic operations on two child nodes.
 public class ALNode extends OpNode {
