+++ RuleEngine/nodes/BooleanNode.java
@@ -1,4 +1,6 @@
+package nodes;
 import java.util.ArrayList;
+import engine.GameState;
 
 public class BooleanNode extends OpNode {
     private String operator;
