+++ RuleEngine/nodes/BooleanOperation.java
@@ -1,3 +1,4 @@
+package nodes;
 public final class BooleanOperation {
     private BooleanOperation() {}
