+++ RuleEngine/nodes/GetNode.java
@@ -1,12 +1,13 @@
+package nodes;
 import java.util.ArrayList;
 import java.util.Map;
+import engine.*;
 
 // Usage: Operand 0 - property name
-//        Operand 1 - source(register or gameobject)
-//        Operand 2 - value
-public class PSetNode extends OpNode {
-    public PSetNode() { super(); }
-    public PSetNode(ArrayList<Node> operands) {
+//        Operand 1 - source (register or gameobject)
+public class GetNode extends OpNode {
+    public GetNode() { super(); }
+    public GetNode(ArrayList<Node> operands) {
         super();
         this.operands.set(0, operands);
     }
@@ -14,9 +15,8 @@ public class PSetNode extends OpNode {
     public LiteralNode execute(GameState currState) {
         LiteralNode e1 = getOperand(0).execute(currState);
         LiteralNode e2 = getOperand(1).execute(currState);
-        LiteralNode e3 = getOperand(2).execute(currState);
 
-        if (e1 == null || e2 == null || e3 == null) {
+        if (e1 == null || e2 == null) {
             System.out.println("Error: Something went wrong processing rset operation");
             return null;
         }
@@ -35,12 +35,8 @@ public class PSetNode extends OpNode {
         } else {
             go = currState.registers.get(str2);
         }
-
-/*      TODO: activate and test after UMass Dining pushes new GameObject definition.
-        if (!go.setTrait(str1, e3.getValue()) {
-            System.out.println("Error: Something went wrong when setting " + str1 + " property of " + str2 + " to " + e3.getValue().toString());
-        }
-*/
+//        TODO: activate and test after UMass Dining pushes new GameObject definition.
+//        return new LiteralNode<Object>(go.getTrait(str1));
         return null;
     }
 }
