+++ RuleEngine/nodes/IfNode.java
@@ -1,4 +1,6 @@
+package nodes;
 import java.util.ArrayList;
+import engine.GameState;
 
 // TODO: This works, but needs more testing.
 //       1. Executing a block when there are multiple rules in that block.
