+++ RuleEngine/nodes/RSetNode.java
@@ -1,16 +1,19 @@
+package nodes;
 import java.util.ArrayList;
 import java.util.Map;
+import engine.*;
 
-// Usage: Operand 0 - property name
-//        Operand 1 - source (register or gameobject)
-public class GetNode extends OpNode {
-    public GetNode() { super(); }
-    public GetNode(ArrayList<Node> operands) {
+// Usage: Operand 0 - target (register name)
+//        Operand 1 - source name (register or gameobject)
+public class RSetNode extends OpNode {
+    public RSetNode() { super(); }
+    public RSetNode(ArrayList<Node> operands) {
         super();
         this.operands.set(0, operands);
     }
 
     public LiteralNode execute(GameState currState) {
+        GameObject go;
         LiteralNode e1 = getOperand(0).execute(currState);
         LiteralNode e2 = getOperand(1).execute(currState);
 
@@ -23,18 +26,15 @@ public class GetNode extends OpNode {
             System.out.println("Error: rset operation only takes strings!");
             return null;
         }
-
         String str1 = (String)e1.getValue();
         String str2 = (String)e2.getValue();
 
-        GameObject go;
         if (str2.charAt(0) == '_') {
             go = currState.findObject(str2.substring(1));
         } else {
             go = currState.registers.get(str2);
         }
-//        TODO: activate and test after UMass Dining pushes new GameObject definition.
-//        return new LiteralNode<Object>(go.getTrait(str1));
+        currState.registers.put(str1, go);
         return null;
     }
 }
