+++ RuleEngine/tests/DemoTest.java
@@ -1,4 +1,7 @@
+package tests;
 import java.util.ArrayList;
+import nodes.*;
+import engine.*;
 
 public class DemoTest {
     public static void main(String[] args) {
