+++ RuleEngine/tests/Test.java
@@ -1,4 +1,7 @@
+package tests;
 import java.util.ArrayList;
+import nodes.*;
+import engine.*;
 
 public class Test {
     public static void main(String[] args) {
